refactor examples module
